Clarify CLI help
